backend/requirements.txt
aiofiles
asyncpg
cachetools
fastapi
httpx
jsonschema
pydantic-settings
sqlalchemy[asyncio]
sse-starlette
uvicorn
